+++ mapsforge-map-awt/src/test/java/org/mapsforge/map/layer/download/tilesource/OpenStreetMapMapnikTest.java
@@ -1,6 +1,7 @@
 /*
  * Copyright 2010, 2011, 2012, 2013 mapsforge.org
  * Copyright 2014 Ludwig M Brinckmann
+ * Copyright 2018 devemux86
  *
  * This program is free software: you can redistribute it and/or modify it under the
  * terms of the GNU Lesser General Public License as published by the Free Software
@@ -34,7 +35,7 @@ public class OpenStreetMapMapnikTest {
         TileSource tileSource = OpenStreetMapMapnik.INSTANCE;
 
         URL tileUrl = tileSource.getTileUrl(new Tile(0, 1, (byte) 2, 256));
-        Assert.assertTrue(tileUrl.toExternalForm().endsWith(".tile.openstreetmap.org:80/2/0/1.png"));
+        Assert.assertTrue(tileUrl.toExternalForm().endsWith(".tile.openstreetmap.org:443/2/0/1.png"));
     }
 
     @Test