+++ mapsforge-map/src/main/java/org/mapsforge/map/layer/download/tilesource/OpenStreetMapMapnik.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2010, 2011, 2012, 2013 mapsforge.org
  * Copyright 2014 Ludwig M Brinckmann
- * Copyright 2014 devemux86
+ * Copyright 2014-2018 devemux86
  *
  * This program is free software: you can redistribute it and/or modify it under the
  * terms of the GNU Lesser General Public License as published by the Free Software
@@ -32,9 +32,9 @@ import java.net.URL;
  */
 public class OpenStreetMapMapnik extends AbstractTileSource {
     public static final OpenStreetMapMapnik INSTANCE = new OpenStreetMapMapnik(new String[]{
-            "a.tile.openstreetmap.org", "b.tile.openstreetmap.org", "c.tile.openstreetmap.org"}, 80);
+            "a.tile.openstreetmap.org", "b.tile.openstreetmap.org", "c.tile.openstreetmap.org"}, 443);
     private static final int PARALLEL_REQUESTS_LIMIT = 8;
-    private static final String PROTOCOL = "http";
+    private static final String PROTOCOL = "https";
     private static final int ZOOM_LEVEL_MAX = 18;
     private static final int ZOOM_LEVEL_MIN = 0;