+++ docs/Changelog.md
@@ -12,6 +12,7 @@
 - Tile sources with api keys [#1028](https://github.com/mapsforge/mapsforge/issues/1028)
 - Render theme fallback internal resources [#1026](https://github.com/mapsforge/mapsforge/issues/1026)
 - MapViewPosition interface [#1044](https://github.com/mapsforge/mapsforge/pull/1044)
+- Android 9 compatibility [#1066](https://github.com/mapsforge/mapsforge/issues/1066)
 - JTS (LocationTech) [#1027](https://github.com/mapsforge/mapsforge/issues/1027)
 - Many other minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/mapsforge/issues?q=is%3Aclosed+milestone%3A0.10.0)